pyproject.toml
[project]
name = "egg-n-bacon-housing"
version = "0.2.0"
description = "Singapore housing data pipeline with ML analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
    "numpy>=1.24.0",
    "seaborn",
    "scikit-learn",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "geopandas",
    "requests",
    "ipywidgets",
    "langchain>=0.3.0",
    "langchain-google-genai==2.0.0",
    "langchain-experimental==0.3.0",
    "langgraph==0.2.32",
    "langgraph-checkpoint==2.0.0",
    "langchain-community",
    "beautifulsoup4",
    "h3==4.1.0b2",
    "boto3",
    "pygwalker",
    "supabase",
    "tabulate==0.9.0",
    "plotly>=6.5.2",
    "cachetools>=6.2.4",
    "statsmodels>=0.14.0",
    "rapidfuzz>=3.14.3",
    "xgboost>=3.1.3",
    "prophet>=1.2.1",
    "libpysal>=4.6.0",
    "esda>=1.5.0",
    "lifelines>=0.27.0",
    "marimo>=0.19.6",
    "scipy>=1.17.0",
    "kaleido>=1.2.0",
    "patsy>=1.0.2",
    "shap>=0.49.1",
    "sf-hamilton>=1.0",
    "tenacity",
    "polars>=1.40.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "pytest>=8.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "coverage[toml]>=7.0.0",
    "ruff>=0.3",
    "libpysal>=4.14.1",
    "esda>=2.8.1",
    "geopandas>=1.1.2",
    "shapely>=2.1.2",
    "kaleido>=1.2.0",
    "playwright>=1.58.0",
    "pre-commit>=3.0.0",
    "mypy>=1.8.0",
]

[tool.hatch.build.targets.sdist]
include = ["src/"]

[tool.hatch.build.targets.wheel]
packages = ["src/egg_n_bacon_housing"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"notebooks/*.py" = ["E402", "E722", "F821", "F841"]
"notebooks/**/*.py" = ["E402", "E722", "F821", "F841"]
"src/egg_n_bacon_housing/analytics/**/*.py" = ["E402", "N806", "N803", "F401", "F821", "F841", "I001"]
"src/egg_n_bacon_housing/components/*.py" = ["N999"]  # Numbered stage files like 01_ingestion.py

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
    "--cov=egg_n_bacon_housing",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-report=xml:coverage.xml",
]

markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, may use external resources)",
    "slow: Slow tests (should be run infrequently)",
    "api: Tests that make API calls (may need mocking)",
]

[tool.coverage.run]
source = ["egg_n_bacon_housing"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
# Mypy type checking configuration
# Strategy: Gradual strict mode - start with basic checking, enable strict as type hints improve
python_version = "3.12"
# strict = true  # TODO: Enable strict mode gradually as type hints are added
ignore_missing_imports = true
warn_unused_configs = true
warn_redundant_casts = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]

# Known issues to address gradually:
# 1. SecretStr fields in config.py - pydantic-settings compatibility with strict mode
# 2. Legacy modules in analytics - need full type annotations

# Override strict mode for legacy/experimental modules
# Priority: Add type hints to these modules gradually, then remove overrides
[[tool.mypy.overrides]]
module = "notebooks.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "src/egg_n_bacon_housing.analytics.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "src/egg_n_bacon_housing.config.*"
disable_error_code = ["assignment"]
ignore_errors = true

[[tool.mypy.overrides]]
module = "src/egg_n_bacon_housing.analytics.*"
ignore_errors = true
